requirements.txt
pygame
PyOpenGL
numpy
noise